bulutapp/requirements.txt
asgiref==3.8.1
Django==3.2.9
lxml==5.2.2
python-docx==1.1.2
pytz==2024.1
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
